Create index.html
This is a current project I am working on for a course I am taking so the code will be updated frequently. I am learning how to implement flexbox and flexgrid as well as JavaScript. Disclaimer: I do not own any rights to the video used in this project however I was required to use an embedded video to pass one of the tests for this project.